server01/flask/requirements.txt
flask==3.0.0
gunicorn==21.2.0
kafka-python==2.0.2
